requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-taggit>=5.0
whitenoise>=6.6
